requirements.txt
crewai
streamlit
python-dotenv
groq


-e .